# docs/swagger.yaml
definitions:
  dto.AddBookRequest:
    properties:
      book_id:
        type: string
    type: object
...
  dto.BaseReviewRequest:
    properties:
      book_id:
        description: |-
          ID книги, к которой относится отзыв (UUID)
          Required: true
          Example: "123e4567-e89b-12d3-a456-426614174000"
        type: string
      rating:
        description: |-
          Оценка (от 1 до 10)
          Required: true
          Example: 9
        maximum: 10
        minimum: 1
        type: integer
      text:
        description: |-
          Текст отзыва
          Required: true
          Example: "Отличная книга, советую!"
        type: string
    required:
    - book_id
    - rating
    - text
    type: object
...
  dto.ReviewResponse:
    description: Ответ API с информацией об отзыве
    properties:
      book_id:
        description: |-
          ID книги, к которой относится отзыв (UUID)
          Required: true
          Example: "123e4567-e89b-12d3-a456-426614174000"
        type: string
      created_at:
        description: |-
          Дата создания
          Example: "2024-02-01T12:00:00Z"
        type: string
      dislikes:
        description: |-
          Дизлайки
          Example: 2
        type: integer
      id:
        description: |-
          Уникальный идентификатор отзыва (ObjectID в MongoDB)
          Example: "60c72b2f5f1b2c001f6f1b20"
        type: string
      likes:
        description: |-
          Лайки
          Example: 10
        type: integer
      rating:
        description: |-
          Оценка (от 1 до 10)
          Required: true
          Example: 9
        maximum: 10
        minimum: 1
        type: integer
      text:
        description: |-
          Текст отзыва
          Required: true
          Example: "Отличная книга, советую!"
        type: string
      updated_at:
        description: |-
          Дата обновления (если есть)
          Example: "2024-02-02T14:30:00Z"
        type: string
      user_id:
        description: |-
          ID автора отзыва (UUID)
          Example: "550e8400-e29b-41d4-a716-446655440000"
        type: string
    required:
    - book_id
    - rating
    - text
    type: object
...
  dto.UpdateReadingProgressRequest:
    properties:
      pages_read:
        type: integer
      status:
        type: string
    type: object
  dto.UserBookResponse:
    properties:
      book_id:
        type: string
      created_at:
        type: string
      pages_read:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.VoteReviewRequest:
    description: Запрос API с голосованием (1 -1 0)
    properties:
      vote:
        description: 1 - лайк, -1 - дизлайк, 0 - удалить голос
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /review/{reviewID}:
    get:
      description: Возвращает карточку отзыва
      parameters:
      - description: ObjectID отзыва
        in: path
        name: reviewID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReviewResponse'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Отзыв не найден
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить отзыв по ID
      tags:
      - Reviews
  /review/{reviewID}/vote:
    post:
      consumes:
      - application/json
      description: Проголосовать за отзыв +1 -1 0
      parameters:
      - description: ObjectID отзыва
        in: path
        name: reviewID
        required: true
        type: string
      - description: Данные для голосования
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/dto.VoteReviewRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
      security:
      - BearerAuth: []
      summary: Проголосовать за отзыв
      tags:
      - Reviews
  /reviews/{reviewID}:
    delete:
      description: Удаляет отзыв по ID (только автор, модератор или админ)
      parameters:
      - description: ObjectID отзыва
        in: path
        name: reviewID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Отзыв успешно удалён'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверные параметры
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет прав на удаление
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Отзыв не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить отзыв
      tags:
      - Reviews
    put:
      consumes:
      - application/json
      description: Редактирует текст и/или оценку отзыва
      parameters:
      - description: ObjectID отзыва
        in: path
        name: reviewID
        required: true
        type: string
      - description: Данные для обновления отзыва
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/dto.BaseReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Отзыв успешно обновлён'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверные параметры
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет прав на редактирование
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Отзыв не найден
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить отзыв
      tags:
      - Reviews
...
  /users/me/books:
    get:
      description: Возвращает список книг, добавленных пользователем
      produces:
      - application/json
      responses:
        "200":
          description: Список книг пользователя
          schema:
            items:
              $ref: '#/definitions/dto.UserBookResponse'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить список книг пользователя
      tags:
      - UserBooks
    post:
      consumes:
      - application/json
      description: Добавляет указанную книгу в список пользователя
      parameters:
      - description: ID книги
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/dto.AddBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: Книга добавлена в список'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Добавить книгу в список пользователя
      tags:
      - UserBooks
  /users/me/books/{bookID}:
    delete:
      description: Удаляет указанную книгу из списка пользователя
      parameters:
      - description: UUID книги
        in: path
        name: bookID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Книга удалена из списка'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить книгу из списка пользователя
      tags:
      - UserBooks
  /users/me/books/{bookID}/progress:
    put:
      consumes:
      - application/json
      description: Обновляет статус и количество прочитанных страниц
      parameters:
      - description: UUID книги
        in: path
        name: bookID
        required: true
        type: string
      - description: Прогресс чтения
        in: body
        name: progress
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateReadingProgressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Прогресс чтения обновлен'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить прогресс чтения
      tags:
      - UserBooks
...
swagger: "2.0"